requirements.txt
bs4
pandas
requests
mysql-connector-python